+++ jgclark.NoteHelpers/src/lib/commands/titleToFilename.js
@@ -7,12 +7,12 @@ import { logDebug, logError, logInfo, logWarn } from '@helpers/dev'
  * Renames the current note to match its title.
  * @returns void
  */
-export function titleToFilename(): void {
+export async function titleToFilename(): Promise<void> {
   try {
     const { note } = Editor
 
     if (note) {
-      renameNoteToTitle(note)
+      const shouldContinue = await renameNoteToTitle(note)
     }
   } catch (error) {
     logError(error)